+++ src/app/user-list/user-list.component.ts
@@ -1,8 +1,10 @@
 import { Component, OnInit } from '@angular/core';
 import { UserService } from '../shared/services/user.service';
+import { DataService } from '../shared/services/data.service';
 import {DataSource} from '@angular/cdk/collections';
 import { Observable } from 'rxjs';
 import {User} from '../shared/models/user.model';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-data',
@@ -14,18 +16,19 @@ export class UserListComponent implements OnInit {
   dataSource = new UserDataSource(this.userService);
   displayedColumns = ['id', 'name', 'email', 'phone',  'company', 'website', 'actions'];
   DeleteUser(id: number) {
-  this.userService.deleteUser(id);
-
+    if ( confirm('sure you want to remove this user ?' )) {
+      this.userService.deleteUser(id);
+    }
   }
-  constructor(private userService: UserService) { }
+  UpdateUser(id: Number) {
+    this.dataService.userid = id;
+    this.router.navigate(['UpdateUser']);
+  }
+
+  constructor(private userService: UserService, private  dataService: DataService, private router: Router) { }
    ngOnInit() {
- }
-
-
-
-
+  }
 }
-
 
 export class UserDataSource extends DataSource<any> {
   constructor(private userService: UserService) {
